Add files via upload
In 2015 a data journalist Walt Hickey analyzed movie ratings data from Fandango Online Movie Rating Forum and found that the rating system was biased and dishonest. Fandango's officials replied that the bias was caused by a bug in the system and promised to fix it. But the actual rating value is not available anymore. The main objective of the project was to determine if Fandango has updated its movie rating system after Hickey's analysis. As we did not have sufficient number of samples for our analysis, we had to change our objective. Our modified objective was If there has been any difference between Fandango's ratings for popular movies released in 2015 and Fandango's ratings for popular movies released in 2016.  For conclusions, please go through the complete project.