Project on 'Intermediate Statistics in Python'
Our objective in this project is to find the two best markets for the advertisement of programming courses on 'Web & Mobile Development', 'Data Science', 'Gaming' etc. of the e-learning company we are working with. For this purpose, we analyzed 2017 New Coder Survey data released by freeCodeCamp, a free e-learning platform that offer courses on Web Development. Please go through the project in order to learn about our findings.