Hypothesis/Chi-square Testing
The project is about winning money in a US television series called "Jeopardy".  In this project we worked with a dataset with old Jeopardy question and answers. We will analyze if studying old questions is going to help us win money in jeopardy. We used "Chi-square Testing" in order to come to a conclusion.